added additional check to entity

+++ src/entities/category.js
@@ -1,5 +1,7 @@
 const Category = require("../models/category");
+const PostEntity = require("./post");
 const Error404 = require("../middleware/error/error404");
+const Error400 = require("../middleware/error/error400");
 
 async function getAllCategories() {
   return Category.query().select();
@@ -13,8 +15,8 @@ async function getCategoryById(id) {
   return Category.query().findById(id);
 }
 
-async function createCategory(post) {
-  return Category.query().insert(post);
+async function createCategory(category) {
+  return Category.query().insert(category);
 }
 
 async function updateCategory(id, data) {
@@ -27,8 +29,14 @@ async function updateCategory(id, data) {
 
 async function removeCategoryById(id) {
   const category = await Category.query().findById(id);
+  const postsByCategory = await PostEntity.getPostsByCategory(id);
   if (!category) {
     throw new Error404();
+  }
+  if (postsByCategory !== 0) {
+    throw new Error400(
+      "You cannot remove the category because there are posts of this category"
+    );
   }
   return Category.query().findById(id).delete();
 }
